+++ changelog.md
@@ -5,11 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### Removed
-- deprecated class `org.tbk.bitcoin.common.util.Hex`
+### Changed
+- upgrade: update bitcoin testcontainer from v23 to v24
 
 ### Fixed
 - fix: use canonical form for property conditions
+
+### Removed
+- deprecated class `org.tbk.bitcoin.common.util.Hex`
 
 ## [0.7.0] - 2023-04-04
 ### Added