+++ changelog.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - upgrade: update bitcoin testcontainer from v23 to v24
 - upgrade: update lightning-kmp from v1.4.1 to v1.4.4
+- upgrade: update lombok from v1.18.20 to v1.18.26
 
 ### Fixed
 - fix: use canonical form for property conditions