+++ changelog.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.8.0] - 2023-04-29
 ### Changed
 - upgrade: update bitcoin testcontainer from v23 to v24
 - upgrade: update lnd testcontainer from v0.16.0-beta to v0.16.1-beta
@@ -171,7 +173,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial release
 
-[Unreleased]: https://github.com/theborakompanioni/bitcoin-spring-boot-starter/compare/0.7.0...HEAD
+[Unreleased]: https://github.com/theborakompanioni/bitcoin-spring-boot-starter/compare/0.8.0...HEAD
+[0.8.0]: https://github.com/theborakompanioni/bitcoin-spring-boot-starter/compare/0.7.0...0.8.0
 [0.7.0]: https://github.com/theborakompanioni/bitcoin-spring-boot-starter/compare/0.6.0...0.7.0
 [0.6.0]: https://github.com/theborakompanioni/bitcoin-spring-boot-starter/compare/0.5.0...0.6.0
 [0.5.0]: https://github.com/theborakompanioni/bitcoin-spring-boot-starter/compare/0.4.1...0.5.0